fsh refactor

+++ professional/ele-32/lab3/convolutional/convolutional.cpp
@@ -19,7 +19,7 @@ int Convolutional::encode(int input) {
 
 void Convolutional::decode(int output) {
 	std::vector<Node> nextStates(this->stateSequence.size(), Node{-1, (int)1e8, -1});
-	for (int i = 0; i < this->stateSequence.size(); i++) {
+	for (int i = 0; i < (int)this->stateSequence.size(); i++) {
 		std::vector<Edge> edges = this->t.getTransitions(i);
 		for (const auto &e : edges) {
 			int cost = this->stateSequence[i].back().cost;
@@ -31,14 +31,14 @@ void Convolutional::decode(int output) {
 		}
 	}
 
-	for (int i = 0; i < this->stateSequence.size(); i++) {
+	for (int i = 0; i < (int)this->stateSequence.size(); i++) {
 		this->stateSequence[i].push_back(nextStates[i]);
 	}
 }
 
 std::vector<int> Convolutional::getSequence() {
 	int minState = 0;
-	for (int i = 0; i < this->stateSequence.size(); i++) {
+	for (int i = 0; i < (int)this->stateSequence.size(); i++) {
 		if (this->stateSequence[i].back().cost < this->stateSequence[minState].back().cost) {
 			minState = i;
 		}
@@ -46,7 +46,7 @@ std::vector<int> Convolutional::getSequence() {
 
 	std::stack<int> s;
 	int currState = minState;
-	for (int i = this->stateSequence[0].size() - 1; i > 0; i--) {
+	for (int i = (int)this->stateSequence[0].size() - 1; i > 0; i--) {
 		s.push(this->stateSequence[currState][i].input);
 		currState = this->stateSequence[currState][i].parent;
 	}
